+++ news/news.rst
@@ -2,6 +2,8 @@
 
 * Add `xpdacq.plan_stubs.mv_and_trigger_and_read` for the multi-detector plan for both XRD and PDF data collection in one run.
 
+* Add `xpdacq.devices.CalibrationData`, a class to store the calibration data of a detector in configuration attributes.
+
 **Changed:**
 
 * <news item>